requirements.txt
asn1crypto==0.24.0
click==7.1.2
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
ffmpeg==1.4
Flask==1.1.2
future==0.18.2
idna==2.6
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.3
opencv-python==4.5.1.48
Pillow==8.0.1
pycrypto==2.6.1
pygobject==3.26.1
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
pyxdg==0.25
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==2.3.1
six==1.15.0
tifffile==2020.9.3
#torch==1.6.0+cu101
#torchvision==0.7.0+cu101
#linux pip install torch==1.6.0+cu101 torchvision==0.7.0+cu101 -f https://download.pytorch.org/whl/torch_stable.html
tqdm==4.54.1
trimesh==3.8.19
Werkzeug==1.0.1
